backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
python-whois==0.9.4
requests==2.31.0
cryptography==42.0.7
scikit-learn==1.4.2
joblib==1.4.0
numpy==1.26.4
Pillow==10.3.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
